refactor(gateway): ファイル構成を変更
