+++ api/internal/gateway/admin/komoju/request/ping.go
